pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpower-gui"
version = "0.2.0"
description = "PySide6 desktop GUI for MCPower Monte Carlo power analysis"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Paweł Lenartowicz", email = "pawellenartowicz@europe.com"}
]
requires-python = ">=3.10"
dependencies = [
    "MCPower==0.5.0",
    "PySide6-Essentials>=6.5.0",
    "pyqtgraph>=0.13.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
mcpower-gui = "mcpower_gui.__main__:main"

[project.gui-scripts]
mcpower-gui = "mcpower_gui.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcpower_gui*"]

[tool.setuptools.package-data]
mcpower_gui = ["acknowledgments.txt", "media/*.gif", "media/*.png", "media/*.ico", "media/*.icns", "*.desktop", "docs/*.md", "*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
